feat: Вынести компонент из story
